# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Property AI Studio (PropVista CRM) API
  version: 1.0.0
  description: 'REST API for PropVista CRM. Base path /api/v1. JWT Bearer auth. Gemini-only
    AI. Roles: customer, agent, admin, super_admin. Guest = no auth. Errors: {error:{code,message,details[]}}.
    Pagination meta. Prices as strings. Notification channels: email, in_app.'
servers:
- url: http://localhost:4001/api/v1
  description: Local development
- url: https://staging-api.propvista.example/api/v1
  description: Staging
- url: https://api.propvista.example/api/v1
  description: Production
...
paths:
...
  /users:
    get:
      tags:
      - Users
      summary: List users
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPublic'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: getUsers
      security:
      - bearerAuth:
        - admin
        - super_admin
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
    post:
      tags:
      - Users
      summary: Create user
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin.'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'VALIDATION_ERROR'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'CONFLICT_DUPLICATE_EMAIL'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'VALIDATION_ERROR'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: postUsers
      security:
      - bearerAuth:
        - admin
        - super_admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
  /users/{id}:
    get:
      tags:
      - Users
      summary: Get user
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin, agent.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'RESOURCE_NOT_FOUND'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: getUserById
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier
    patch:
      tags:
      - Users
      summary: Update user
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'RESOURCE_NOT_FOUND'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'VALIDATION_ERROR'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: patchUserById
      security:
      - bearerAuth:
        - admin
        - super_admin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
    delete:
      tags:
      - Users
      summary: Delete user
      description: 'Auth: JWT Bearer. Authorization: super_admin.'
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'RESOURCE_NOT_FOUND'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: deleteUserById
      security:
      - bearerAuth:
        - super_admin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
              description: Stable application error code (SRS Appendix P)
              enum:
              - AUTH_INVALID_CREDENTIALS
              - AUTH_TOKEN_EXPIRED
              - AUTH_FORBIDDEN
              - VALIDATION_ERROR
              - RESOURCE_NOT_FOUND
              - CONFLICT_DUPLICATE_EMAIL
              - LEAD_IDEMPOTENCY_REPLAY
              - AI_UNAVAILABLE
              - AI_TIMEOUT
              - BULK_PARSE_ERROR
              - BULK_VALIDATION_FAILED
              - PROPERTY_PUBLISH_BLOCKED
              - RATE_LIMITED
              - INTERNAL_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
    Meta:
      type: object
      required:
      - page
      - pageSize
      - total
      - totalPages
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    UserRole:
      type: string
      enum:
      - customer
      - agent
      - admin
      - super_admin
...
    UserPublic:
      type: object
      required:
      - id
      - email
      - role
      - isActive
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required:
      - email
      - password
      - role
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        fullName:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
          description: Deactivate or reactivate without soft-deleting
...
security:
- bearerAuth: []